requirements.txt
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flask[async]==3.0.2
aiohttp==3.10.3
python-dotenv==1.0.1
requests==2.33.1
WTForms==3.2.1
